chore: remove legacy api
